requirements.txt
anyio==3.6.1
black==22.6.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.79.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.2
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2